Fixes to runner
