+++ bout_runners/database/database_utils.py
@@ -14,116 +14,6 @@ from bout_runners.utils.subprocesses_functions import run_subprocess
 from bout_runners.executor.runner_utils import run_settings_run
 
 
-def obtain_project_parameters(settings_path):
-    """
-    Return the project parameters from the settings file.
-
-    Parameters
-    ----------
-    settings_path : Path
-        Path to the settings file
-
-    Returns
-    -------
-    run_parameters_dict : dict
-        Dictionary containing the parameters given in BOUT.settings
-        On the form
-        >>> {'section': {'parameter': 'value'}}
-
-    Notes
-    -----
-    1. The sectionless part of BOUT.settings will be renamed `global`
-    2. In the `global` section, the keys `d` and the directory to the
-       BOUT.inp file will be removed
-    3. If the section `all` is present in BOUT.settings, the section
-       will be renamed `all_boundaries` as `all` is a protected SQL
-       keyword
-    4. The section `run` will be dropped due to bout_runners own
-       `run` table
-    """
-    # The settings file lacks a header for the global parameter
-    # Therefore, we add add the header [global]
-    with settings_path.open('r') as settings_file:
-        settings_memory = f'[global]\n{settings_file.read()}'
-
-    config = configparser.ConfigParser()
-    config.read_string(settings_memory)
-
-    parameter_dict = dict()
-
-    for section in config.sections():
-        parameter_dict[section] = dict()
-        for key, val in config[section].items():
-            # Strip comments
-            capture_all_but_comment = '^([^#]*)'
-            matches = re.findall(capture_all_but_comment, val, re.M)
-
-            # Exclude comment line
-            if len(matches) == 0:
-                continue
-
-            # Capitalize in case of boolean
-            stripped_val = matches[0].capitalize()
-
-            # If type is not found, type is str
-            try:
-                val = ast.literal_eval(stripped_val)
-            except (SyntaxError, ValueError):
-                val = stripped_val
-
-            parameter_dict[section][key] = val
-
-    # NOTE: Bug in .settings: -d path is captured with # not in use
-    bout_inp_dir = settings_path.parent
-    parameter_dict['global'].pop('d', None)
-    parameter_dict['global'].pop(str(bout_inp_dir).lower(), None)
-
-    if 'all' in parameter_dict.keys():
-        parameter_dict['all_boundaries'] = parameter_dict.pop('all')
-
-    # Drop run as bout_runners will make its own table with that name
-    parameter_dict.pop('run', None)
-
-    return parameter_dict
-
-
-def cast_parameters_to_sql_type(parameter_dict):
-    """
-    Return the project parameters from the settings file.
-
-    Parameters
-    ----------
-    parameter_dict : dict
-        Dictionary containing the parameters given in BOUT.settings
-        On the form
-        >>> {'section': {'parameter': 'value'}}
-
-    Returns
-    -------
-    parameter_dict_sql_types : dict
-        Dictionary containing the parameters given in BOUT.settings
-        On the form
-        >>> {'section': {'parameter': 'value_type'}}
-    """
-    type_map = {'bool': 'INTEGER',  # No bool type in SQLite
-                'float': 'REAL',
-                'int': 'INTEGER',
-                'str': 'TEXT'}
-
-    parameter_dict_sql_types = parameter_dict.copy()
-
-    for section in parameter_dict.keys():
-        for key, val in parameter_dict[section].items():
-            # If type is not found, type is str
-            try:
-                val_type = type(ast.literal_eval(val))
-            except (SyntaxError, ValueError):
-                val_type = str
-
-            parameter_dict[section][key] = type_map[val_type.__name__]
-
-    return parameter_dict_sql_types
-
 
 def get_system_info_as_sql_type():
     """
